Multiple additions
Migrated to foundation 5.0
Fixed foundation dropdowns to show with custom styling
Fixed bug when closing an EmptyView (isDestroyed method issue) Now we check if there is a model AND if that model has a function isDestroyed. For some reason the view would close and suddenly have this empty backbone model.
Edited topbar colors, back to the deep black.
Fixed delete button by adding an extra-tiny class. Had to because foundation 5.0 migration. Looked fine in foundation 4.0.
Fixed a bug where an item would be deleted and wouldn't show the red wrapper.